BE-thuvien-main/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.7.0
django-extensions==4.1
Pillow==11.3.0
PyMySQL==1.1.1